Fix contrail-common/base unit tests
1. Added popen and pclose wrappers for Windows to address_util_test
2. Added util_test.cc to flake tests - it received 3 independent time
   measures and compared its absolute value to be lower than 1ms.
   Sometimes it's bigger than 1ms even tough it's still correct behaviour

Partial Jira Bug: WIND-183

+++ base/index_map.h
@@ -80,7 +80,7 @@ public:
         ValueType *value = values_[index];
         values_[index] = NULL;
         delete value;
-        for (ssize_t i = values_.size() - 1; i >= 0; i--) {
+        for (int64_t i = values_.size() - 1; i >= 0; i--) {
             if (values_[i] != NULL) {
                 break;
             }
